+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Aggregates are now stored on multiple files with the `Filesystem` adapter
 - (Optional) Entities id column with the `SQL` adapter now use the Aggregate id as a value, the columns in the Aggregate column referencing these columns have been removed
+- `Formal\ORM\Raw\Aggregate\Collection::properties()` has been renamed `::entities()`
 
 ## 1.2.0 - 2024-01-15